Add ONIE password - for Secure Boot Extended
This defaults to setting an empty root password if
SECURE_BOOT_EXT is set.

The machine install-platform file can override the set_default_passwd
function to set a password on a per-machine basis.

The default root password for the KVM virtual build is 'onie'

Tested with SECURE_BOOT_ENABLE on, and SECURE_BOOT_EXT on and off.

Only for secure boot move /etc/passwd to be stored on hdd. Add default
 password for mellanox machine Flags for Onie password enable:
 SECURE_BOOT_ENABLE or SECURE_BOOT_EXT

Signed-off-by: Andriy Dobush <andriyd@nvidia.com>
Signed-off-by: Alex Doyle <adoyle@nvidia.com>