Secure Grub - enabled
This adds the SECURE_GRUB option to change
GRUB's behavior in the following ways:

It now has a user account and password, specified
by GRUB_USER, and GRUB_PASSWD_PLAINTEXT, or
GRUB_PASSWD_HASH

It uses the gpg-sign.sh script to create detached gpg
signatures to use in verifying files that it loads.
This includes configuration, the kernel, etc

Note that modifying the configuration file can render
the system unbootable.
So...don't.

Signed-off-by: Andriy Dobush <andriyd@nvidia.com>
Signed-off-by: Alex Doyle <adoyle@nvidia.com>