Toggle password support based on secure boot
This adds -open and -secured versions of
/bin/onie-console ( the login script)
and
/etc/passwd
The open-system startup script will detect
if the Secure Boot UEFI variable is set or not.
If set, it sets softlinks to
/bin/onie-console-secure
/mnt/onie-boot/onie-config/etc/passwd
(unless this is a recovery iso, in which case it
does nothing )
If not set, links go to
/bin/onie-console-open
/etc/passwd-open

The resulting behavior is that, if Secure Boot is set,
the user is prompted for a login/password.
If not, hittng 'enter' brings the user to a prompt.

Signed-off-by: Michael Shych <michaelsh@nvidia.com>
Signed-off-by: Alex Doyle <adoyle@nvidia.com>